feat(traverse): implement `DoubleEndedIterator` for sibling iterators
Oops, forgot to implement these earlier.

Getting the siblings is ok for normal nodes, but for root nodes it
doesn't really work, since root nodes don't have siblings. Thus, if a
root node is detected whilst attempting to build sibling iterators, the
tail will be `None`, meaning it will be an iterator of a
single element - itself.